amazon-dashboard/dashboard-v1/backend/requirements.txt
# FastAPI and web server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.0

# Authentication
python-jose[cryptography]>=3.3.0

# Amazon SP-API
python-amazon-sp-api>=1.0.0

# Configuration
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# HTTP client
httpx>=0.26.0

# Development
black>=23.0.0
ruff>=0.1.0
mypy>=1.0.0
pytest>=7.0.0
pytest-asyncio>=0.23.0
